+++ framework/Source/GPUImageKuwaharaFilter.h
@@ -8,6 +8,6 @@
 }
 
 /// The radius to sample from when creating the brush-stroke effect, with a default of 3. The larger the radius, the slower the filter.
-@property(readwrite, nonatomic) GLuint radius;
+@property(readwrite, nonatomic) NSUInteger radius;
 
 @end
