+++ framework/Source/GPUImageWhiteBalanceFilter.h
@@ -9,9 +9,9 @@
     GLint temperatureUniform, tintUniform;
 }
 //choose color temperature, in degrees Kelvin
-@property(readwrite, nonatomic) int temperature;
+@property(readwrite, nonatomic) CGFloat temperature;
 
 //adjust tint to compensate
-@property(readwrite, nonatomic) int tint;
+@property(readwrite, nonatomic) CGFloat tint;
 
 @end
